Extract the WSG traectory generator (and other LCM related items) into their own module.
Also stop using a rigid body tree to initialize the WSG LCM classes.
